requirements.txt
requests==2.32.3
pymongo==4.8.0
python-dotenv==1.0.1
openai==1.46.0
scikit-learn==1.5.1
tqdm==4.66.5
sentence-transformers==2.5.1
huggingface-hub==0.25.2
transformers>=4.37.0
torch>=2.1.0
python-docx==1.1.0
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0
streamlit==1.39.0
openpyxl==3.1.5
html2image==2.0.5




# requests==2.32.3
# pymongo==4.8.0
# python-dotenv==1.0.1
# openai==1.46.0
# scikit-learn==1.5.1
# tqdm==4.66.5
# sentence-transformers==2.5.1
# torch
# tqdm

# python-docx
# numpy<2

# huggingface-hub==0.19.4

# fastapi
# uvicorn
# pydantic
# requests
